+++ CHANGES.rst
@@ -8,6 +8,7 @@ Version 0.2
 not yet released
 
 - Return nothing when parsing irrelevant HTML table rows.
+- Return extracted ranks as a generator instead of a list.
 
 
 Version 0.1